backend/requirements.txt
# Web框架
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
gunicorn==20.1.0
Werkzeug==2.2.3

# 数据库
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.5.post1
alembic==1.10.2

# 工具库
python-dotenv==1.0.0
pydantic==1.10.6
Pillow==9.4.0
requests==2.28.2
pytest==7.2.2
black==23.1.0
flake8==6.0.0

# AI模型API
openai==0.27.0
google-generativeai==0.3.1
anthropic==0.5.0
